:sparkles: CHORE: Add searchTasks action for searching tasks by query
This commit adds the searchTasks action, which allows users to search for tasks based on a query string. The action utilizes the OpenAI API to generate an embedding for the query and then performs a vector search to find similar tasks. The search results are returned as an array of task objects.