:sparkles: CHORE: Add AddProjectDialog component for creating new projects
This commit adds the AddProjectDialog component, which allows users to create new projects. The component includes a form for entering the project name and a button to submit the form. Upon submission, the createAProject mutation is called to create the project in the database. If the project is successfully created, a success toast is displayed and the user is redirected to the project page.