:rocket: REFACTOR: Update ConvexAdapter to exclude "id" field when creating a user, which was causing issue in deploy.
The ConvexAdapter in ConvexAdapter.ts has been updated to exclude the "id" field when creating a user. This change was made to align with the changes in the createUser mutation in authAdapter.ts, where the "id" field is no longer needed. By removing the "id" field from the createUser function in ConvexAdapter, we ensure that the user object passed to the mutation does not contain an "id" property.