:boom: CHORE: Update RootLayout to include session and Providers moved up from loggedin layout to root.
This commit updates the RootLayout component in layout.tsx to include the session object obtained from the auth function. It also wraps the children components with the Providers component, which provides necessary context for the application. This change ensures that the session is available throughout the layout and its child components.